backend/requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
